Refactors the controllers
